Feat : Dog skin model 추가
- manet 추가
- 실험 위한 변경 사항 추가

+++ src/train.py
@@ -14,7 +14,7 @@ def train(config):
     optimizer = exp.optimizer
     loss = exp.loss
     scheduler = exp.scheduler
-    trainer = exp.trainer(logger=exp.logger, callbacks=exp.callbacks)
+    trainer = exp.trainer(logger=exp.logger, precision=16, callbacks=exp.callbacks)
 
     model = exp.module(
         model=architecture,
